Update referee.py
The user should be able to propose a solution using less power plants than the given ones.
Meaning that `powers` should be a sublist of `plants`. The easiest way to check that is comparing list's counters.
Then eventually, we can later propose random tests with more than the minimal number of power plants, since it seems complicated to me.